requirements.txt
aioredis==1.0.0
asgi-redis==1.4.3
asgiref==2.2.0
async-timeout==2.0.0
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
certifi==2017.11.5
channels==2.0.2
channels-redis==2.1.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.1.0
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.0.3
django-allauth==0.34.0
django-cors-headers==2.1.0
django-jet==1.0.7
django-rest-auth==0.9.3
django-rest-swagger==2.1.2
djangorestframework==3.7.3
google-api-python-client==1.4.1
gunicorn==19.7.1
hiredis==0.2.0
httplib2==0.10.3
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
msgpack==0.5.6
msgpack-python==0.5.1
oauth2client==4.1.2
oauthlib==2.0.6
openapi-codec==1.3.2
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
python3-openid==3.1.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
static3==0.7.0
Twisted==17.9.0
txaio==2.9.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
zope.interface==4.4.3
